+++ api/models/EventOrder.js
@@ -30,7 +30,7 @@ module.exports = {
       type: Sequelize.ENUM("NEW", "RECEIVED", "REQUESTED", "SUBMITTED", "PAID", "PROCESSING", "CANCELLED", "SHIPPED", "DELIVERED", "COMPLETED"),
       defaultValue: 'NEW',
     },
-		
+
 		token: {
       type: Sequelize.STRING(32),
 			unique: true,
@@ -95,6 +95,19 @@ module.exports = {
 			}
 		},
 
+    ItemNameArray: {
+      type: Sequelize.VIRTUAL,
+      get: function () {
+        try {
+          const thisEvent = this.getDataValue('Event');
+          const event = thisEvent ? [thisEvent.title] : [];
+          return event;
+        } catch (e) {
+          sails.log.error(e);
+        }
+      }
+    },
+
 	},
 	associations: () => {
 		EventOrder.belongsTo(User);
@@ -105,7 +118,7 @@ module.exports = {
 		classMethods: {
 			findByIdHasJoin: async(id) => {
 				try {
-					return await RecipeOrder.findOne({
+					return await EventOrder.findOne({
 						where: {
 							id
 						},
@@ -118,7 +131,7 @@ module.exports = {
 			},
 			deleteById: async(id) => {
 				try {
-					return await RecipeOrder.destroy({
+					return await EventOrder.destroy({
 						where: {
 							id
 						}